pyproject.toml
[project]
name = "b3_app"
version = "0.1.0"
description = ""
requires-python = ">=3.11"
dependencies = [
    "django>=5.1.3",
    "django-ninja>=1.3.0",
    "pika>=1.3.2",
    "requests>=2.32.3",
    "ruff>=0.7.4",
    "django-ninja-extra>=0.21.6",
]

[tool.ruff.lint]
select = [
    # pycodestyle
    "E",
    # Pyflakes
    "F",
    # pyupgrade
    "UP",
    # flake8-bugbear
    "B",
    # flake8-simplify
    "SIM",
    # isort
    "I",
]
